(shmif/tui) wrong type for hint, resize on wndhint
TUI clients has previously blindly accepted whatever sizes it gets,
and only using hints for subwindows in order to indicate capacity
for popups and so on.

This patch prepares for constraints on main segment as well, being
tested through extending trayicon tool with handling text- inputs
such as various bar formats.

FossilOrigin-Name: ee4a0752e102381206e642c7d4a4d540d54653e883c56b904b07a92f5a46a262

+++ src/shmif/arcan_shmif_event.h
@@ -1451,7 +1451,7 @@ enum ARCAN_TARGET_SKIPMODE {
  */
 		struct {
 			int32_t x, y;
-			uint16_t w, h;
+			uint32_t w, h;
 			uint32_t parent;
 			uint8_t border[4];
 			uint8_t edge;
